moves cube, sphere, bunny to entity directory, refactor FileInternalDataProvider
