+++ stlib/entities/cube.py
@@ -10,10 +10,8 @@ class CubeParameters(EntityParameters):
     deformable : bool = False
 
     def model_post_init(self, _):
-        # TODO : 
-        # 1. apply size as scale in geometry, material, collision and visual 
         if self.deformable:
-            self.geometry = FileParameters(filename="mesh/sphere.vtk", elementType=ElementType.TETRAHEDRA)
+            self.geometry = FileParameters(filename="mesh/sphere.vtk", elementType=ElementType.TETRAHEDRA, scale=self.size)
             self.material = DeformableMaterialParameters()
         return
